trabalhando no perfil

+++ server/src/controllers/userControllers.js
@@ -95,11 +95,11 @@ async function update(req,res,next) {
 
     if(req.files.avatar){
         av = req.files.avatar[0]
-        newData.avatar = av.path
+        newData.avatar = `http://localhost:3001/image/${av.filename}`
     }
     if(req.files.background){
         bg = req.files.background[0]
-        newData.background = bg.path
+        newData.background = `http://localhost:3001/image/${bg.filename}`;
     }
 
     try{
@@ -110,13 +110,13 @@ async function update(req,res,next) {
         if (!((userId == user.id) || (userRole == "admin"))) throw new createHttpError(403, "You don't have permission to do this");
 
         if (newData.avatar){
-            if (user.avatar) {
+            if (user.avatar && user.avatar != "undefined") {
                 fs.unlinkSync(user.avatar)
             }
         }
 
         if (newData.background){
-            if (user.background) {
+            if (user.background && user.avatar != "undefined") {
                 fs.unlinkSync(user.background)
             }
         }
